match_predictor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.2.1
requests==2.31.0
python-dotenv==1.0.0
lightgbm==4.1.0
scikit-learn==1.4.1.post1
joblib==1.3.2
numpy>=1.24.0
scipy>=1.12.0
pydantic>=2.4.2
